requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
databases[asyncpg]==0.9.0
asyncpg==0.29.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-dotenv==1.0.1
httpx==0.27.0
pydantic==2.7.1
pydantic-settings==2.2.1
python-multipart==0.0.9
